% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR_fsts.R
\name{calc.fst.wright}
\alias{calc.fst.wright}
\title{Calculate Fst using Wright (1943)'s formulation ((ht-hs)/ht)}
\usage{
calc.fst.wright(al1, al2)
}
\arguments{
\item{al1}{A table of allele frequencies for pop 1}

\item{al2}{A table of allele frequencies for pop 2}
}
\value{
data.frame containing the columns:
 Hs1 = expected heterozygosity in population 1 (vcf1)
 Hs2 = expected heterozygosity in population 2 (vcf2)
 Hs = weighted average expected heterozygosity within populations
 Ht = expected heterozygosity among populations
 Fst= Fst
 NumAlleles=The number of alleles (2 for biallelic loci)
 Num1=number of individuals genotyped & included in Fst calculations in population 1 (vcf1)
 Num2=number of individuals genotyped & included in Fst calculations in population 2 (vcf2)
 @export
}
\description{
Calculate Fst using Wright (1943)'s formulation ((ht-hs)/ht)
}
