% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR_popgen.R
\name{calc.het}
\alias{calc.het}
\title{calculate observed heterozygosity}
\usage{
calc.het(vcf.row, pop.list = NULL)
}
\arguments{
\item{vcf.row}{A row of a vcf file, used in conjunction with apply}

\item{pop.list}{A list of population strings to search for to subset vcf (pop names must be in individual ID names)}
}
\value{
Observed heterozygosity
}
\description{
calculate observed heterozygosity
}
\note{
proportion of diploid genotypes in the sample that are heterozygotes (Hohenlohe et al 2010)
}
\examples{
vcf.file<-system.file("extdata", "example.vcf.txt",package = "gwscaR")
vcf<-parse.vcf(vcf.file)
all.het<-apply(vcf,1,calc.het)
}
