% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR_popgen.R
\name{calc.pi}
\alias{calc.pi}
\title{sliding window pi and rho - across all snps}
\usage{
calc.pi(vcf.row)
}
\arguments{
\item{vcf.row}{A row of a vcf file; use this in conjunction with apply}
}
\value{
pi The nucleotide diversity at that site
}
\description{
sliding window pi and rho - across all snps
}
\note{
pi = 1-sum((ni choose 2)/(n choose i)); ni is number of alleles i in sample, n = sum(ni)

Jones et al. (2012) used 2500bp sliding windows with a step size 500bp<-more than just SNPs, but I'll just focus on SNPs

Hohenlohe did a similar thing and weighted pi by all nt sites (not just SNPs) but rho by SNPs only
}
