% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR_popgen.R
\name{calc.rho}
\alias{calc.rho}
\title{Calculate rho (private alleles)}
\usage{
calc.rho(vcf.row, pop.list)
}
\arguments{
\item{vcf.row}{A row of a data frame containing vcf-formatted data}

\item{pop.list}{A list with the population names (these should be contained within the sample names)}
}
\description{
Calculate rho (private alleles)
}
\note{
rho=1 if allele in pop j is only found in that pop and at least one ind was genotyped at that site in each pop; rho = 0 otherwise
}
