% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR.R
\name{changeorder.df}
\alias{changeorder.df}
\title{A function to reorder a data.frame}
\usage{
changeorder.df(dat, order.list)
}
\arguments{
\item{dat}{A data.frame. Group ID should be in column 1}

\item{order.list}{A list of factors for reordering}
}
\value{
dat.new The reordered dataframe
}
\description{
A function to reorder a data.frame
}

