% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR.R
\name{choose.one.snp}
\alias{choose.one.snp}
\title{Choose one SNP per RAD locus from a vcf}
\usage{
choose.one.snp(vcf)
}
\arguments{
\item{vcf}{A data.frame in vcf format}
}
\value{
new.vcf A data.frame with only one SNP per RAD locus
}
\description{
Choose one SNP per RAD locus from a vcf
}

