% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR.R
\name{combine.vcfs}
\alias{combine.vcfs}
\title{A function to merge two vcf data frames}
\usage{
combine.vcfs(vcf1, vcf2, vcf.name = "merge.vcf")
}
\arguments{
\item{vcf1}{A dataframe with genomic data in vcf format}

\item{vcf2}{A dataframe with genomic data in vcf format}

\item{vcf.name}{A name for a file with the new vcf file (default is merge.vcf)}
}
\value{
vcf A new data.frame in vcf format
}
\description{
A function to merge two vcf data frames
}
