% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR.R
\name{extract.gt.vcf}
\alias{extract.gt.vcf}
\title{A function to pull out only the genotype fields from vcf-format data}
\usage{
extract.gt.vcf(vcf)
}
\arguments{
\item{vcf}{A data.frame with genotype data in vcf format}
}
\value{
vcf A new vcf with only the GT data
}
\description{
A function to pull out only the genotype fields from vcf-format data
}
