% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR_popgen.R
\name{fst.ibd.byloc}
\alias{fst.ibd.byloc}
\title{Calculate isolation by distance per locus}
\usage{
fst.ibd.byloc(ped.file, dist.mat, pop.order)
}
\arguments{
\item{ped.file}{A data.frame in ped format}

\item{dist.mat}{A distance matrix containing the geeographic distances between populations}

\item{pop.order}{A list with the order for the populations}
}
\value{
results.mantel A data.frame with two columns containing the mantel test results
}
\description{
Calculate isolation by distance per locus
}
