% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR_fsts.R
\name{fst.one.plink}
\alias{fst.one.plink}
\title{Calculate pairwise fsts from a dataset in ped format}
\usage{
fst.one.plink(raw, group1, group2, cov.thresh = 0.2, loc_names = NULL)
}
\arguments{
\item{raw}{A data.frame with data in ped format}

\item{group1}{A list with the individuals in group 1}

\item{group2}{A list with the individuals in group 2}

\item{cov.thresh}{A threshold for the number of individuals in the populations (default is 0.2)}
}
\value{
fst.dat A data.frame with columns:
 Locus Name = the locus name
 Hs1 = expected heterozygosity in pop 1
 Hs2 = expected heterozygosity in pop 2
 Hs = expected heterozygosity within populations
 Ht = expected heterozygosity among populations
 Fst = Fst
 NumAlleles = number of alleles at the locus
 Num1 = the number of individuals in pop 1
 Num2 = the number of individuals in pop 2
}
\description{
Calculate pairwise fsts from a dataset in ped format
}
