% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR.R
\name{fst.one.vcf}
\alias{fst.one.vcf}
\title{Calculate fsts from a single vcf}
\usage{
fst.one.vcf(vcf, group1, group2, cov.thresh = 0.2, maf = 0.05)
}
\arguments{
\item{vcf}{A data.frame in vcf format}

\item{group1}{A list of individual names for pop1}

\item{group2}{A list of individual names for pop2}

\item{cov.thresh}{A proportion of individuals required for genotyping individuals (default is 0.2)}

\item{maf}{A minimum allele frequency (default is 0.05)}
}
\value{
out A data.frame with columns:
Hs1 (expected heterozygosity in pop 1), Hs2 (expected heterozygosity in pop2),
Hs (expected heterozygosity within pops), Ht (expected heterozygosity between pops),
Fst, NumAlleles, Num1 (number of individuals in pop1), and Num2 (number of individuals in pop2)
}
\description{
Calculate fsts from a single vcf
}

