% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR.R
\name{fst.plot.rect}
\alias{fst.plot.rect}
\title{Plot genome-wide statistics from a data frame with rectangles.}
\usage{
fst.plot.rect(fst.dat, ci.dat = NULL, sig.col = c("red", "yellow"),
  pt.col = "grey7", fst.name = "Fst", chrom.name = "Chrom",
  bp.name = "BP", axis.size = 0.5, scaffold.order = NULL, groups = NULL,
  print.names = FALSE, y.lim = NULL, group.boundaries = NULL,
  pt.cex = 0.5, ...)
}
\arguments{
\item{fst.dat}{The data.frame containing at least three columns: the statistic to be plotted, the chromosome ID, and the BP ID for each locus. Each row is a locus.}

\item{ci.dat}{A vector containing two values, upper and lower cutoff values (in that order).
Points above or below the cutoffs will be colored based on sig.col. Default is NULL, which turns this option off.}

\item{sig.col}{A vector containing two color values, the first for points above the upper cutoff value and the second for points below the lower cutoff value.
The defaults are red and yellow.}

\item{pt.col}{The color of the points. The default is grey7.}

\item{fst.name}{The name of the column containing the statistic to be plotted. Default is "Fst".}

\item{chrom.name}{The name of the column containing the chromosome information for each locus. Default is "Chrom".}

\item{bp.name}{The name of the column containing the basepair information for each locus. Default is "BP".}

\item{axis.size}{The value of cex.axis passed to the plotting of the y-axis. Default is 0.5. Set axis.size = 0 to suppress plotting of the y-axis.}

\item{scaffold.order}{A vector or list containing the names of the chromosomes/scaffolds in the order in which you would like them to be plotted.}

\item{groups}{A vector indicating which chromosomes to plot (generally to exclude those scaffolds not found in this particular set of statistics due to pruning/filters)}

\item{print.names}{A TRUE/FALSE value indicating whether chromosome/scaffold IDs should be printed beneath the x-axis.}

\item{y.lim}{The limits for the y-axis.}

\item{group.boundaries}{A data.frame with first column naming the chromosomes/linkage groups, second column having the last number of loci on the chromosome for indexing purposes.
e.g.:
 Chrom   End
 LG1     3945850
 LG2     435215
This parameter is used to make multiple plots from the same overall dataset have the same widths.}
}
\value{
xes The fst.dat data frame with new values in BP, scaled to be the sequence in which points are plotted based on their position on the chromosome and the order the chromosomes are plotted in.
}
\description{
Plot genome-wide statistics from a data frame with rectangles.
}
\note{
This fst.plot function is more efficient but this will work for most cases (might be buggy in some cases)
}
\seealso{
Flanagan & Jones 2017
}
