% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR_popgen.R
\name{fst.pst.byloc}
\alias{fst.pst.byloc}
\title{Compare Pst and Fst locus by locus from a ped file}
\usage{
fst.pst.byloc(ped.file, trait.df, pop.order, trait.ind)
}
\arguments{
\item{ped.file}{A data.frame with data in ped format}

\item{trait.df}{A data.frame containing all of the traits data}

\item{pop.order}{An order in which the populations should be analyzed}

\item{trait.ind}{The trait index in the trait.df}
}
\value{
A dataframe containing the restuls of the mantel test (one column for observations and one for p-values)
}
\description{
Compare Pst and Fst locus by locus from a ped file
}
