% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR.R
\name{fst.sig}
\alias{fst.sig}
\title{Find significant loci using chi-square test}
\usage{
fst.sig(fst.df)
}
\arguments{
\item{fst.df}{A data.frame containing at least the following columns: Fst, Num1 (number of individuals genotyped in pop 1), Num2 (number of individuals genotyped in pop2)}
}
\value{
The same dataframe but with additional columns
}
\description{
Find significant loci using chi-square test
}
