% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR_popgen.R
\name{get.nj}
\alias{get.nj}
\title{Generate a distance matrix for a SNP using fsts}
\usage{
get.nj(vcf.row, pop.list)
}
\arguments{
\item{vcf.row}{A row of a vcf file}

\item{pop.list}{A list of populations in the order you want them to appear in the matrix}
}
\value{
fst.tree A neighbor-joining tree from the ape package
}
\description{
Generate a distance matrix for a SNP using fsts
}
\examples{
fst.trees<-list()
for(vcf.row in 1: nrow(vcf)){
  fst.trees<-c(fst.trees,get.nj(vcf[vcf.row,],pop.list)) #getting an error
}
}
