% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR.R
\name{gwsca}
\alias{gwsca}
\title{Conduct selection components analysis}
\usage{
gwsca(vcf, locus.info, group1, group2, prop.ind.thresh = 0.5,
  maf.cutoff = 0.05)
}
\arguments{
\item{vcf}{A data.frame with genotype data in vcf format}

\item{locus.info}{A list of column names with the locus info (e.g. c(#CHROM,POS))}

\item{group1}{A list of column names with individuals from the first group}

\item{group2}{A list of column names with individuals from the second group}

\item{prop.ind.thresh}{A proportion of indidivudals requried in each population (default is 0.5)}

\item{maf.cutoff}{A minimum allele frequency cutoff (default is 0.05)}
}
\value{
A dataframe with the Fst values and Chi-squared values
}
\description{
Conduct selection components analysis
}
