% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR_plot.R
\name{gwsca.vioplot}
\alias{gwsca.vioplot}
\title{Create violin plots with}
\usage{
gwsca.vioplot(x, ..., range = 1.5, h = NULL, ylim = NULL,
  names = NULL, horizontal = FALSE, col = "magenta",
  border = "black", lty = 1, lwd = 1, rectCol = "black",
  colMed = "white", pchMed = 19, at, add = FALSE, wex = 1,
  drawRect = TRUE, plot.axes = TRUE, axis.box = FALSE,
  plot.ann = TRUE)
}
\arguments{
\item{...}{legend parameters}
}
\value{
Optionally returns the summary statistics of the data.
}
\description{
Create violin plots with
}
