% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR_plot.R
\name{outer.legend}
\alias{outer.legend}
\title{Add a legend to the margins of a multiplot figure}
\usage{
outer.legend(...)
}
\arguments{
\item{...}{legend parameters}
}
\value{
Nothing is returned
}
\description{
Add a legend to the margins of a multiplot figure
}
