% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR.R
\name{pairwise.fst}
\alias{pairwise.fst}
\title{Calculate pairwise Fst values}
\usage{
pairwise.fst(ped, allele1, allele2, pop.order)
}
\arguments{
\item{ped}{A dataframe with data in ped format, where each allele for each locus is in its own column}

\item{allele1}{An index for the first allele at the locus}

\item{allele2}{An index for the second allele at the locus}

\item{pop.order}{A list with the Pop IDs in the correct order}
}
\value{
dat.var A matrix of pairwise fst values (calculated as (ht-hs)/ht)
}
\description{
Calculate pairwise Fst values
}
