% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR_popgen.R
\name{pairwise.pst}
\alias{pairwise.pst}
\title{Calculate pairwise Pst between population pairs}
\usage{
pairwise.pst(dat, pop.order)
}
\arguments{
\item{dat}{A dataframe with the trait values, first column must be the pop ID}

\item{pop.order}{A list of the order of the populations}
}
\value{
dat.var A data.frame with the pairwise Pst values
}
\description{
Calculate pairwise Pst between population pairs
}
