% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR.R
\name{parse.vcf}
\alias{parse.vcf}
\title{Read in a vcf file}
\usage{
parse.vcf(filename)
}
\arguments{
\item{filename}{The name of the vcf file}
}
\value{
a dataframe containing the contents of the vcf file, including headers.
}
\description{
Read in a vcf file
}
\examples{
vcf<-parse.vcf("example.vcf")
}
\seealso{
Flanagan & Jones 2017
}

