% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR.R
\name{plotting.structure}
\alias{plotting.structure}
\title{This plots the output of structure}
\usage{
plotting.structure(structure.out, k, pop.order, filename = paste("str.k", k,
  ".jpeg", sep = ""), make.file = TRUE, plot.new = TRUE, colors = NULL,
  xlabel = TRUE, ylabel = NULL)
}
\arguments{
\item{structure.out}{A data.frame containing the structure output.}

\item{k}{The value of K used in Structure}

\item{pop.order}{The order in which the populations are plotted.}

\item{filename}{The name of the jpeg for the output file.}

\item{make.file}{A boolean (TRUE/FALSE) indicating whether the jpeg should be created}

\item{plot.new}{A boolean (TRUE/FALSE) indicating whether this is being added to an existing plot}

\item{colors}{A list of colors for the structure colors (if not provided, defaults to rainbow palette)}

\item{xlabel}{A boolean (TRUE/FALSE) indicating whether x labels should be plotted (default is TRUE)}

\item{ylabel}{An optional label for the y-axis.}
}
\description{
This plots the output of structure
}

