% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR_sim.R
\name{popgen.sim}
\alias{popgen.sim}
\title{Simulate genetic data for populations with different allele frequencies}
\usage{
popgen.sim(npops = 8, inds = 10, nloc = 10000, p = rep(0.01, 8),
  outname = "simulated.ped", analyze = TRUE, ns = c(12, 24, 36, 48,
  96))
}
\arguments{
\item{npops}{The number of populations to simulate}

\item{inds}{The number of individuals per population to simulate}

\item{nloc}{The number of loci to simulate}

\item{p}{A vector of average minor allele frequencies per locus to simulate}
}
\value{
A data.frame with ped data
}
\description{
Simulate genetic data for populations with different allele frequencies
}
