% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR.R
\name{pst.mantel}
\alias{pst.mantel}
\title{Calulate pairwise Psts and test for IBD}
\usage{
pst.mantel(trait.df, comp.df, id.index, pop.order)
}
\arguments{
\item{trait.df}{A data frame with trait values for all traits and all individuals}

\item{comp.df}{A dataframe with the distance values}

\item{id.index}{The index value for the trait being calculated}

\item{pop.order}{A vector of population names in the order you want them.}
}
\value{
results.mantel A data.frame with the results of the mantel test
}
\description{
Calulate pairwise Psts and test for IBD
}

