% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR.R
\name{sem}
\alias{sem}
\title{Calculate the standard error of the mean}
\usage{
sem(x)
}
\arguments{
\item{x}{A list of values}
}
\value{
sem A value of the standard error of the mean
}
\description{
Calculate the standard error of the mean
}

