% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR_popgen.R
\name{sliding.avg}
\alias{sliding.avg}
\title{Calculate a sliding average}
\usage{
sliding.avg(win.start, dat, width)
}
\arguments{
\item{win.start}{The beginning of where you want to start your sliding window}

\item{dat}{A vector (or row of a data frame) containing the values you would like to calculate your sliding window over.}

\item{width}{The width of the sliding window}
}
\value{
The average value of dat within that sliding window.
}
\description{
Calculate a sliding average
}
