% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR_popgen.R
\name{sliding.window}
\alias{sliding.window}
\title{Calculate a sliding window}
\usage{
sliding.window(vcf, chr, stat = "pi", width = 250, pop.list = NULL,
  nsteps = 50)
}
\arguments{
\item{vcf}{A vcf file}

\item{chr}{The chromosome you want to calculate the sliding window in}

\item{stat}{the population genetics statistic you want to use. choices are "pi" or "rho"}

\item{width}{The width for the sliding window. Default is 250}

\item{pop.list}{An optional parameter allowing you to specificy the populations you want to include.}

\item{nsteps}{The number of steps in the sliding window}
}
\value{
A vector with the averaged statistic
}
\description{
Calculate a sliding window
}
