% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR_utility.R
\name{std.by.mean}
\alias{std.by.mean}
\title{Standardize a variable by its mean}
\usage{
std.by.mean(x)
}
\arguments{
\item{x}{A vector of numeric values}
}
\value{
newx A vector of the input values after standardizing to the mean.
}
\description{
Standardize a variable by its mean
}
\note{
Useful for generate Bayenv2 environmental variables
}
