% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR_plot.R
\name{treemix.cov.plot}
\alias{treemix.cov.plot}
\title{Plot covariance matrix from treemix}
\usage{
treemix.cov.plot(stem, poporder, colors = c("blue", "yellow", "red"),
  ...)
}
\arguments{
\item{stem}{The filename basic stem for that run}

\item{poporder}{The list of populations in the order to plot them.}
}
\value{
cp The covariance matrix in the plotting order.
}
\description{
Plot covariance matrix from treemix
}
