% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR_popgen.R
\name{treemix.from.vcf}
\alias{treemix.from.vcf}
\title{Generate a treemix file from vcf}
\usage{
treemix.from.vcf(vcf, pop.list)
}
\arguments{
\item{vcf}{A vcf file}

\item{pop.list}{a list of populations}
}
\value{
A matrix of allele counts for each population for each locus, aka an input file for treemix
}
\description{
Generate a treemix file from vcf
}
