% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR.R
\name{vcf.alleles}
\alias{vcf.alleles}
\title{Extract alleles from a vcf row}
\usage{
vcf.alleles(vcf.row)
}
\arguments{
\item{vcf.row}{A vcf row (containing only the individuals and locus info)}
}
\value{
al1 A list of alleles
}
\description{
Extract alleles from a vcf row
}
