% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR.R
\name{vcf.cov.ind}
\alias{vcf.cov.ind}
\title{Calculate per-individual coverage from a vcf file}
\usage{
vcf.cov.ind(vcf)
}
\arguments{
\item{vcf}{A data.frame containing data in vcf format.}
}
\value{
icov A data.frame with one row for each individual and columns:
   NumMissing: The number of missing loci
   NumPresent: The number of loci genotyped in this individual
   AvgCovRef: The average coverage for reference alleles across all genotyped loci in this individual
   AvgCovAlt: The average coverage for alternative alleles across all genotyped loci in this individual
   AvgCovTot: The average total coverage (ref + alt coverage) across all genotyped loci in this individual
   PropHet: Proportion of loci at which this individual is heterozygous
   NumReads: The total number of reads for this individual
}
\description{
Calculate per-individual coverage from a vcf file
}
