% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR.R
\name{vcf.cov.loc}
\alias{vcf.cov.loc}
\title{Calculate per-locus coverage from a vcf file}
\usage{
vcf.cov.loc(vcf, subset = NULL)
}
\arguments{
\item{vcf}{A data.frame containing data in vcf format.}

\item{subset}{A list of the column names of the individuals to be used (optional)}
}
\value{
cov.dat A data.frame with one row for each locus and 14 columns:
   Chrom: The chromosome
   Pos: The position/BP on the chromosome
   Locus: The Locus ID
   NumMissing: The number of individuals not gentoyped at this locus
   NumPresent: The number of individuals genotyped at this locus
   PropMissing: The proportion of individuals not genotyped at this locus
   AvgCovRef: The average coverage of the reference allele in genotyped individuals
   AvgCovAlt: The average coverage of the alternative allele in genotyped individuals
   AvgCovRatio: The average ration of Reference/Alternative allele coverage
   AvgCovTotal: The average of the number of reference + alternative reads per individual
   CovVariance: The variance in coverage among individuals
   NumHet: The number of individuals genotyped as heterozygotes
   PropHet: The proportion of individuals genotyped as heterozygotes
   TotalNumReads: The total number of reads at this locus
}
\description{
Calculate per-locus coverage from a vcf file
}

