% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR_utility.R
\name{vcf2coanAF}
\alias{vcf2coanAF}
\title{Convert a vcf df to a coancestry input file}
\usage{
vcf2coanAF(vcf, out.name = "coancestry_afs.txt")
}
\arguments{
\item{vcf}{A data.frame in vcf format}

\item{out.name}{A name for a an output file (default is coancestry_afs.txt)}
}
\value{
co.afs A dataframe with all of the allele frequencies
}
\description{
Convert a vcf df to a coancestry input file
}
