% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR_utility.R
\name{vcf2coanGT}
\alias{vcf2coanGT}
\title{Convert a vcf df to a coancestry input file}
\usage{
vcf2coanGT(vcf, out.name = "coancestry_gty.txt")
}
\arguments{
\item{vcf}{A data.frame in vcf format}

\item{out.name}{A name for a an output file (default is coancestry_gty.txt)}
}
\value{
co.gt A dataframe with all of the genotypes
}
\description{
Convert a vcf df to a coancestry input file
}
\examples{
vcf.file<-system.file("extdata", "example.vcf.txt",package = "gwscaR")
vcf<-parse.vcf(vcf.file)
co.gts<-vcf2coanGT(vcf,"coancestry_gty.txt")
}
