% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR_utility.R
\name{vcf2gpop}
\alias{vcf2gpop}
\title{Convert a vcf df to a genepop file}
\usage{
vcf2gpop(vcf, pop.list = NULL, pop.map = NULL, gpop.name)
}
\arguments{
\item{vcf}{A data.frame in vcf format}

\item{pop.list}{A list of population names (the individuals must have names containing the population names)}

\item{pop.map}{A data frame with col 1 = ID names and col 2 = pop names}

\item{gpop.name}{A name for the output genepop file}
}
\value{
gpop A dataframe in genepop format
}
\description{
Convert a vcf df to a genepop file
}
