% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR.R
\name{calc.afs.vcf}
\alias{calc.afs.vcf}
\title{Calculate allele frequencies from a vcf file}
\usage{
calc.afs.vcf(vcf.row)
}
\arguments{
\item{vcf.row}{A single row of a vcf (used in conjunctioin with apply)}
}
\value{
A data.frame with the columns: Chrom,Pos,Ref,RefFreq,Alt,AltFreq
}
\description{
Calculate allele frequencies from a vcf file
}
