% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR.R
\name{calc.afs.vcf}
\alias{calc.afs.vcf}
\title{A function that calculates the allele frequencies from a vcf}
\usage{
calc.afs.vcf(vcf)
}
\arguments{
\item{vcf}{A data.frame with genotype data in vcf format}
}
\value{
out A data.frame with columns Chrom, Pos, Ref (the reference allele),
RefFreq (frequency of reference allele), Alt (alternative allele), and AltFreq (frequency of the alternative allele)
}
\description{
A function that calculates the allele frequencies from a vcf
}
