% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwscaR_sim.R
\name{pcadapt_subset}
\alias{pcadapt_subset}
\title{Subset a ped file and run pcadapt on it}
\usage{
pcadapt_subset(ped, num_inds, subname)
}
\arguments{
\item{ped}{The ped data.frame or file}

\item{inds}{The number of individuals per population to simulate}

\item{nloc}{The number of loci to simulate}

\item{p}{A vector of average minor allele frequencies per locus to simulate}
}
\value{
A data.frame with ped data
}
\description{
Subset a ped file and run pcadapt on it
}
